feat: 💄 基本完成五育分析模块奖项设置管理功能的前后端并修复了若干bug
- 在五育分析相关表中添加了school_id字段（修复了记录无关联学校导致非superAdmin无法获取数据的bug）。
- 在合适字段添加了索引，如记录创建时间字段。
- 去除了五育分析相关表中的主键id自增功能，采用随机id，并将类型修改为BIGINT。

+++ ads-admin/src/main/java/io/ads/modules/analysis/entity/AwardRecordsEntity.java
@@ -18,7 +18,7 @@ public class AwardRecordsEntity {
     /**
      * 主键
      */
-	private Integer id;
+	private Long id;
     /**
      * 学生学号
      */
@@ -26,7 +26,7 @@ public class AwardRecordsEntity {
     /**
      * 奖项id
      */
-	private Integer awardId;
+	private Long awardId;
     /**
      * 获奖等级
      */
